backend/requirements.txt
# Core Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-Limiter==3.8.0

# AI & Processing
openai>=1.58.1
httpx>=0.27.0,<0.28.0

# Excel Processing
openpyxl==3.1.5

# PDF Processing Libraries (Multiple options for compatibility)
pdfplumber==0.11.4
PyPDF2==3.0.1
PyMuPDF==1.24.0

# Statistical Analysis
numpy>=1.26.0

# Database (PostgreSQL + ORM)
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2-binary==2.9.9

# HTTP requests (for Supabase Auth API verification)
requests>=2.31.0